Build and use static libc++ in MacOS.
We are already using static libstdc++ in Linux due to a problem
now appearing in MacOS when using Boost.Test.

When shared C++ library (with C++ interface) is used and
executable does not export its new/delete operators, a situation
of one program/library allocating objects with "new" operator
and another program/library deallocating them with "delete"
operator makes program to crash.

We haven't yet seen these crashes in Windows, even using
Boost.Test. That may be pure luck or because Microsoft std C++
library is coded in a way that makes it not happen.